+++ x-pack/plugins/file_upload/kibana.json
@@ -1,0 +1,8 @@
+{
+  "id": "fileUpload",
+  "version": "8.0.0",
+  "kibanaVersion": "kibana",
+  "server": true,
+  "ui": false,
+  "requiredPlugins": ["usageCollection"]
+}
