+++ x-pack/plugins/file_upload/server/import_data.ts
@@ -5,19 +5,20 @@
  */
 
 import { IScopedClusterClient } from 'kibana/server';
-import { INDEX_META_DATA_CREATED_BY } from '../../../common/constants/file_datavisualizer';
+import { INDEX_META_DATA_CREATED_BY } from '../common/constants';
 import {
   ImportResponse,
   ImportFailure,
   Settings,
   Mappings,
   IngestPipelineWrapper,
-} from '../../../common/types/file_datavisualizer';
-import { InputData } from './file_data_visualizer';
+} from '../common';
+
+export type InputData = any[];
 
 export function importDataProvider({ asCurrentUser }: IScopedClusterClient) {
   async function importData(
-    id: string,
+    id: string | undefined,
     index: string,
     settings: Settings,
     mappings: Mappings,
@@ -77,7 +78,7 @@ export function importDataProvider({ asCurrentUser }: IScopedClusterClient) {
     } catch (error) {
       return {
         success: false,
-        id,
+        id: id!,
         index: createdIndex,
         pipelineId: createdPipelineId,
         error: error.body !== undefined ? error.body : error,
