+++ x-pack/plugins/maps_file_upload/server/routes/file_upload.js
@@ -9,7 +9,7 @@ import { updateTelemetry } from '../telemetry/telemetry';
 import { MAX_BYTES } from '../../common/constants/file_import';
 import { schema } from '@kbn/config-schema';
 
-export const IMPORT_ROUTE = '/api/fileupload/import';
+export const IMPORT_ROUTE = '/api/maps/fileupload/import';
 
 export const querySchema = schema.maybe(
   schema.object({
