+++ x-pack/plugins/ml/common/types/file_datavisualizer.ts
@@ -68,52 +68,3 @@ export interface FindFileStructureResponse {
   timestamp_field?: string;
   should_trim_fields?: boolean;
 }
-
-export interface ImportResponse {
-  success: boolean;
-  id: string;
-  index?: string;
-  pipelineId?: string;
-  docCount: number;
-  failures: ImportFailure[];
-  error?: any;
-  ingestError?: boolean;
-}
-
-export interface ImportFailure {
-  item: number;
-  reason: string;
-  doc: ImportDoc;
-}
-
-export interface Doc {
-  message: string;
-}
-
-export type ImportDoc = Doc | string;
-
-export interface Settings {
-  pipeline?: string;
-  index: string;
-  body: any[];
-  [key: string]: any;
-}
-
-export interface Mappings {
-  _meta?: {
-    created_by: string;
-  };
-  properties: {
-    [key: string]: any;
-  };
-}
-
-export interface IngestPipelineWrapper {
-  id: string;
-  pipeline: IngestPipeline;
-}
-
-export interface IngestPipeline {
-  description: string;
-  processors: any[];
-}
