+++ x-pack/plugins/ml/kibana.json
@@ -10,8 +10,8 @@
     "data",
     "cloud",
     "features",
+    "fileUpload",
     "licensing",
-    "usageCollection",
     "share",
     "embeddable",
     "uiActions",
