+++ x-pack/plugins/ml/public/application/datavisualizer/file_based/components/combined_fields/utils.ts
@@ -8,11 +8,8 @@ import { i18n } from '@kbn/i18n';
 import { cloneDeep } from 'lodash';
 import uuid from 'uuid/v4';
 import { CombinedField } from './types';
-import {
-  FindFileStructureResponse,
-  IngestPipeline,
-  Mappings,
-} from '../../../../../../common/types/file_datavisualizer';
+import { FindFileStructureResponse } from '../../../../../../common/types/file_datavisualizer';
+import { IngestPipeline, Mappings } from '../../../../../../../file_upload/common';
 
 const COMMON_LAT_NAMES = ['latitude', 'lat'];
 const COMMON_LON_NAMES = ['longitude', 'long', 'lon'];
