+++ x-pack/plugins/ml/server/routes/apidoc.json
@@ -43,7 +43,6 @@
 
     "FileDataVisualizer",
     "AnalyzeFile",
-    "ImportFile",
 
     "ResultsService",
     "GetAnomaliesTableData",
