+++ x-pack/plugins/ml/server/routes/schemas/file_data_visualizer_schema.ts
@@ -24,20 +24,3 @@ export const analyzeFileQuerySchema = schema.maybe(
     timestamp_format: schema.maybe(schema.string()),
   })
 );
-
-export const importFileQuerySchema = schema.object({
-  id: schema.maybe(schema.string()),
-});
-
-export const importFileBodySchema = schema.object({
-  index: schema.maybe(schema.string()),
-  data: schema.arrayOf(schema.any()),
-  settings: schema.maybe(schema.any()),
-  /** Mappings */
-  mappings: schema.any(),
-  /** Ingest pipeline definition */
-  ingestPipeline: schema.object({
-    id: schema.maybe(schema.string()),
-    pipeline: schema.maybe(schema.any()),
-  }),
-});
