+++ x-pack/tsconfig.json
@@ -17,6 +17,7 @@
     "plugins/global_search_providers/**/*",
     "plugins/graph/**/*",
     "plugins/features/**/*",
+    "plugins/file_upload/**/*",
     "plugins/embeddable_enhanced/**/*",
     "plugins/event_log/**/*",
     "plugins/enterprise_search/**/*",
@@ -103,6 +104,7 @@
     { "path": "./plugins/enterprise_search/tsconfig.json" },
     { "path": "./plugins/event_log/tsconfig.json" },
     { "path": "./plugins/features/tsconfig.json" },
+    { "path": "./plugins/file_upload/tsconfig.json" },
     { "path": "./plugins/global_search_bar/tsconfig.json" },
     { "path": "./plugins/global_search_providers/tsconfig.json" },
     { "path": "./plugins/global_search/tsconfig.json" },
