+++ x-pack/tsconfig.refs.json
@@ -16,6 +16,7 @@
     { "path": "./plugins/enterprise_search/tsconfig.json" },
     { "path": "./plugins/event_log/tsconfig.json" },
     { "path": "./plugins/features/tsconfig.json" },
+    { "path": "./plugins/file_upload/tsconfig.json" },
     { "path": "./plugins/global_search_bar/tsconfig.json" },
     { "path": "./plugins/global_search_providers/tsconfig.json" },
     { "path": "./plugins/global_search/tsconfig.json" },
